server controller function segregation
