+++ server/controllers/orderControllers.js
@@ -1,0 +1,98 @@
+
+const Orders = require("../models/orders")
+const multer  = require('multer')
+const storage = multer.diskStorage({
+  destination: function (req, file, cb) {
+    cb(null, '../client/src/uploads')
+  },
+  filename: function (req, file, cb) {
+    cb(null, req.body.senderId + '_'+ Math.ceil(Math.random()*123232) + '.' + file.mimetype.split('/')[1])
+  }
+})
+
+const upload = multer({ storage: storage })
+
+
+const PostOrder= upload.single('orderImage'), async (req, res) => {
+    try {
+      const updatedFields = {...req.body, receiverCoordinates: JSON.parse(req.body?.receiverCoordinates),senderCoordinates: JSON.parse(req.body?.senderCoordinates),ordersImageName: req.file.filename}
+      const newOrders = new Orders(updatedFields);
+      await newOrders.save().then(data => {
+        if (data) {
+          res.status(200).json({ message: "Your Order is on the way" })
+        }
+      })
+    } catch (err) {
+      console.log("err" + err)
+      res.status(500).json({ message: err })
+    }
+  }
+
+  const GetOrder = async (req, res) => {
+    try {
+      const skipStartPages = req.query.size * (req.query.page -1)
+      let totalItem = await Orders.find().count()
+      if(totalItem % req.query.size != 0){
+        totalItem = Math.ceil((totalItem / req.query.size))
+      }else{
+        totalItem = (totalItem / req.query.size) 
+      }
+      //we find all the orders for that particular user who requested the orders list 
+      const ordersList = await Orders.find().skip(skipStartPages).limit(req.query.size)
+      res.json({ordersList,totalItem})
+    } catch (err) {
+      console.log(err)
+      res.status(500).json({ message: err })
+    }
+  }
+  const GetOrderSenderID = async (req, res) => {
+    try {
+      const skipStartPages = req.query.size * (req.query.page -1)
+      let totalItem = await Orders.find().count()
+      if(totalItem % req.query.size != 0){
+        totalItem = Math.ceil((totalItem / req.query.size))
+      }else{
+        totalItem = (totalItem / req.query.size) 
+      }
+    
+      //we find all the orders for that particular user who requested the orders list 
+      const ordersList = await Orders.find({ senderId: req.params.senderId }).skip(skipStartPages).limit(req.query.size)
+      res.json({ordersList,totalItem})
+    } catch (err) {
+      res.status(500).json({ message: err })
+    }
+  }
+
+  const PutOrder = async (req, res) => {
+    try {
+      //we find all the orders for that particular user who requested the orders list 
+      const updateData = await Orders.findByIdAndUpdate(req.body._id, req.body)
+      if (updateData) {
+        res.status(200).json({
+          message: 'updated data'
+        })
+      }
+    } catch (err) {
+      res.status(500).json({ message: err })
+    }
+  }
+
+  const DeleteOrder = async (req, res) => {
+    try{
+      const deleteOrder = await Orders.findByIdAndDelete(req.params.id)
+      if(deleteOrder){
+        res.json({
+          message: "Order deleted"
+        })
+      }
+    }catch(err){
+      console.log(err)
+    }
+  
+  }
+
+  exports.GetOrder = GetOrder;
+  exports.PutOrder = PutOrder;
+  exports.DeleteOrder = DeleteOrder;
+  exports.PostOrder = PostOrder;
+  exports.GetOrderSenderID = GetOrderSenderID;
